Only test .js files in test/dialects
When enumerating dialect test files, the current algorithm attempts to
require() all files.  This is a problem for editors which create
swap/backup/temporary files and on systems where thumbnailers or other
system programs create files outside the control of the user.

Although we could attempt to exclude each of these files specifically,
or perform some exclusions based on hidden file names or attributes, it
seems easier to just limit the search to .js files, since that is the
only extension of interest currently.

Signed-off-by: Kevin Locke <klocke@quantpost.com>