Support INSERT with SELECT
Add support for inserting the result of a query in addition to inserting
literal values.  Update Insert to only append to valueSet when
appropriate.  Update the Postgres dialect to only output "VALUES" as
part of an INSERT statement if values are present (including default
values).  Add tests.

Note:  There is some ambiguity of the user intent when insert is called
without any values.  The user may intend to either insert a row
consisting entirely of default values or they may intend to insert a row
based on a subsequently added SELECT statement.  Although we could defer
resolution until the SELECT statement is added (or not) through either
manipulating the query object from the select method or additional logic
in the dialect, the approach taken in this commit is to use a simple
heuristic:  If the insert includes any columns, it is assumed that a
SELECT will follow, otherwise default values are used.

This commit addresses issue #111.

Signed-off-by: Kevin Locke <klocke@quantpost.com>