Clone Column returned by .value()
Previously, calling .value() would mutate the Column object.  This
causes unexpected behavior when subsequent use of the column carries the
previously set value.  For example:

table.insert(col1.value(1));
table.insert(col1).select(col2).from(table2);

Would cause a syntax error because insert(col1) would be interpreted as
inserting the value 1 into col1 and would not accept an appended select
statement.

This is a change which could break any users depending on this behavior.
Although none of the tests depend on this behavior and the new behavior
seems more reasonable (to me, obviously), there is risk of user
breakage.

Signed-off-by: Kevin Locke <klocke@quantpost.com>