Provide mechanism for SELECTing literal values
Previously, the API did not provide a way to include a literal value in
a SELECT statement.  Although users could include an un-escaped string
in the generated statement, this would require users to perform their
own value quoting and escaping.  Although there was previously not much
use for literal values in SELECT, it is not an uncommon use when the
SELECT is part of an INSERT statement.

Add a .value() method to the Sql object which allows users to create
ParameterNodes for values.  Users can then pass the result to .select()
to get literal values in their SELECT statements safely.

Add a test for this use case.

Signed-off-by: Kevin Locke <klocke@quantpost.com>