feat(types): add functionCallCreator typing
The functionCallCreator method isn't new in any way but was missing in the
typings.

Obviously to implement this it was necessary to add the FunctionCallNode type
which just exposes the ValueExpression and AliasNode mixin. Since other Nodes
us it as well I thought about separating it out but I couldn't figure out how
to tell TS that it returns typeof self. I found [this in the
handbook](https://www.typescriptlang.org/docs/handbook/generics.html#using-class-types-in-generics)
which looks like kind of what I want but couldn't get it to work in this case.